OnsetSegmentation: refactor into feature curve + detector
