+++ include/algorithms/public/OnsetSegmentation.hpp
@@ -10,11 +10,11 @@ under the European Union’s Horizon 2020 research and innovation programme
 
 #pragma once
 
+#include "OnsetDetectionFunctions.hpp"
 #include "WindowFuncs.hpp"
 #include "../util/FFT.hpp"
 #include "../util/FluidEigenMappings.hpp"
 #include "../util/MedianFilter.hpp"
-#include "../util/OnsetDetectionFuncs.hpp"
 #include "../../data/FluidIndex.hpp"
 #include "../../data/TensorTypes.hpp"
 #include <Eigen/Eigen>
@@ -31,61 +31,21 @@ public:
   using ArrayXd = Eigen::ArrayXd;
   using ArrayXcd = Eigen::ArrayXcd;
 
-  OnsetSegmentation(index maxSize) : mFFT(maxSize), mWindowStorage(maxSize) {}
+  OnsetSegmentation(index maxSize) : mODF{maxSize} {}
 
   void init(index windowSize, index fftSize, index filterSize)
   {
-    makeWindow(windowSize);
-    prevFrame = ArrayXcd::Zero(fftSize / 2 + 1);
-    prevPrevFrame = ArrayXcd::Zero(fftSize / 2 + 1);
-    mFilter.init(filterSize);
-    mFFT.resize(fftSize);
+    mODF.init(windowSize, fftSize, filterSize);
     mDebounceCount = 1;
-    mPrevFuncVal = 0;
-    mInitialized = true;
-  }
-
-  void makeWindow(index windowSize)
-  {
-    mWindowStorage.setZero();
-    WindowFuncs::map()[mWindowType](windowSize, mWindowStorage);
-    mWindow = mWindowStorage.segment(0, windowSize);
-    mWindowSize = windowSize;
   }
 
   double processFrame(RealVectorView input, index function, index filterSize,
                       double threshold, index debounce = 0,
                       index frameDelta = 0)
   {
-    assert(mInitialized);
-    ArrayXd in = _impl::asEigen<Eigen::Array>(input);
-    double  funcVal = 0;
-    double  filteredFuncVal = 0;
-    double  detected = 0.;
-    if (filterSize >= 3 &&
-        (!mFilter.initialized() || filterSize != mFilter.size()))
-      mFilter.init(filterSize);
-
-    ArrayXcd frame = mFFT.process(in.segment(0, mWindowSize) * mWindow);
-    auto     odf = static_cast<OnsetDetectionFuncs::ODF>(function);
-    if (function > 1 && function < 5 && frameDelta != 0)
-    {
-      ArrayXcd frame2 =
-          mFFT.process(in.segment(frameDelta, mWindowSize) * mWindow);
-      funcVal = OnsetDetectionFuncs::map()[odf](frame2, frame, frame);
-    }
-    else
-    {
-      funcVal =
-          OnsetDetectionFuncs::map()[odf](frame, prevFrame, prevPrevFrame);
-    }
-    if (filterSize >= 3)
-      filteredFuncVal = funcVal - mFilter.processSample(funcVal);
-    else
-      filteredFuncVal = funcVal - mPrevFuncVal;
-
-    prevPrevFrame = prevFrame;
-    prevFrame = frame;
+    double filteredFuncVal =
+        mODF.processFrame(input, function, filterSize, frameDelta);
+    double detected{0};
 
     if (filteredFuncVal > threshold && mPrevFuncVal < threshold &&
         mDebounceCount == 0)
@@ -102,18 +62,9 @@ public:
   }
 
 private:
-  using WindowTypes = WindowFuncs::WindowTypes;
-  FFT          mFFT{1024};
-  ArrayXd      mWindowStorage;
-  ArrayXd      mWindow;
-  index        mWindowSize{1024};
-  index        mDebounceCount{1};
-  ArrayXcd     prevFrame;
-  ArrayXcd     prevPrevFrame;
-  double       mPrevFuncVal{0.0};
-  WindowTypes  mWindowType{WindowTypes::kHann};
-  MedianFilter mFilter;
-  bool         mInitialized{false};
+  index                   mDebounceCount{1};
+  OnsetDetectionFunctions mODF;
+  double                  mPrevFuncVal{0.0};
 };
 
 } // namespace algorithm
